+++ src/com/iesam/domain/models/Alumno.java
@@ -1,56 +1,12 @@
 package com.iesam.domain.models;
 
-public class Alumno {
-    private String id;
-    private String nombre;
-    private String apellidos;
-    private String dni;
-    private String direccion;
+public class Alumno extends Persona{
     private String madre;
     private String padre;
     private String email;
 
 
     //Getters and Setters
-    public String getId() {
-        return id;
-    }
-
-    public void setId(String id) {
-        this.id = id;
-    }
-
-    public String getNombre() {
-        return nombre;
-    }
-
-    public void setNombre(String nombre) {
-        this.nombre = nombre;
-    }
-
-    public String getApellidos() {
-        return apellidos;
-    }
-
-    public void setApellidos(String apellidos) {
-        this.apellidos = apellidos;
-    }
-
-    public String getDni() {
-        return dni;
-    }
-
-    public void setDni(String dni) {
-        this.dni = dni;
-    }
-
-    public String getDireccion() {
-        return direccion;
-    }
-
-    public void setDireccion(String direccion) {
-        this.direccion = direccion;
-    }
 
     public String getMadre() {
         return madre;
